refactor: Improve internal structure enable UI flexibility
