+++ .changeset/beige-things-speak.md
@@ -1,0 +1,5 @@
+---
+'playroom': patch
+---
+
+refactor: Improve internal structure enable UI flexibility
