+++ lib/makeWebpackConfig.js
@@ -26,9 +26,9 @@ module.exports = async (playroomConfig, options) => {
   const ourConfig = {
     mode: options.production ? 'production' : 'development',
     entry: {
-      index: [require.resolve('../src/index.tsx')],
-      frame: [require.resolve('../src/frame.tsx')],
-      preview: [require.resolve('../src/preview.tsx')],
+      index: [require.resolve('../src/entries/index.tsx')],
+      frame: [require.resolve('../src/entries/frame.tsx')],
+      preview: [require.resolve('../src/entries/preview.tsx')],
     },
     output: {
       filename: '[name].[contenthash].js',
