+++ src/entries/index.tsx
@@ -1,14 +1,13 @@
-import faviconInvertedPath from '../images/favicon-inverted.png';
-import faviconPath from '../images/favicon.png';
-
-import Playroom, { type PlayroomProps } from './Playroom/Playroom';
-import { StoreProvider } from './StoreContext/StoreContext';
-import playroomComponents from './components';
-import playroomConfig from './config';
-import { renderElement } from './render';
-import playroomSnippets from './snippets';
-import playroomThemes from './themes';
-import { hmrAccept } from './utils/hmr';
+import faviconInvertedPath from '../../images/favicon-inverted.png';
+import faviconPath from '../../images/favicon.png';
+import Playroom, { type PlayroomProps } from '../Playroom/Playroom';
+import { StoreProvider } from '../StoreContext/StoreContext';
+import playroomComponents from '../components';
+import playroomConfig from '../config';
+import { renderElement } from '../render';
+import playroomSnippets from '../snippets';
+import playroomThemes from '../themes';
+import { hmrAccept } from '../utils/hmr';
 
 const suppliedWidths = playroomConfig.widths || [320, 375, 768, 1024];
 const widths: PlayroomProps['widths'] = [...suppliedWidths, 'Fit to window'];
